Add files via upload
The Expense Tracker Application features a user-friendly interface built with Tkinter and connects to an SQLite database to manage user expenses. The application enables users to add categories for expenses, input individual expenses, and view their spending history. It automatically creates necessary database tables to ensure that data is organized and accessible. Users can input details such as the date, amount, and category of each expense, allowing for insightful analysis of their financial habits. The application is a practical tool for anyone looking to manage their finances more effectively.